W4-D3, Hau Phan : Fix comment bt7

+++ mvn-hauphan/w4-d3/migration_validation/db/migrate/20220516072054_add_members_count_to_teams.rb
@@ -1,5 +1,5 @@
 class AddMembersCountToTeams < ActiveRecord::Migration[7.0]
   def change
-    add_column :teams, :members_count, :decimal
+    add_column :teams, :members_count, :integer
   end
 end
